+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.49] - 2023-11-24
+### Changed
+- Fixed memory issue in internal downloader
+- Changed unpack process to handle cancels
+
 ## [2.0.48] - 2023-11-15
 ### Changed
 - Reverted dockerfile again as the packagemanager still doesn't have .net 8.